pyproject.toml
[tool.poetry]
name = "mkdocs-extraenv-plugin"
version = "0.1.0"
description = "An MkDocs plugin to define extra configuration from environment variables"
authors = ["Jacques Supcik <jacques.supcik@hefr.ch>"]
repository = "https://github.com/supcik/mkdocs-extraenv-plugin"
license = "Apache-2"
readme = "README.md"
packages = [{ include = "extraenv" }]
keywords = ["mkdocs", "python", "markdown", "wiki"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology"
]
exclude = ["test"]

[tool.poetry.dependencies]
python = "^3.10"
mkdocs = "^1.4.3"
python-dotenv = "^1.0.0"
mkdocs-material = { version = "^9.1.15", optional = true }
mkdocs-macros-plugin = { version = "^1.0.1", optional = true }

[tool.poetry.extras]
test = ["mkdocs-material", "mkdocs-macros-plugin"]

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.plugins]

[tool.poetry.plugins."mkdocs.plugins"]
extraenv = "extraenv.plugin:ExtraEnvPlugin"
